requirements.txt
lxml
beautifulsoup4
mysql-connector-python